% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmCDM.R
\name{apmCDM}
\alias{apmCDM}
\title{Fit an Partial-Mastery Additive Cognitive Diagnosis Model (PM-CDM).}
\usage{
apmCDM(data, q, Qmatrix, control = list(), start.par = NULL, ...)
}
\arguments{
\item{data}{Matrix \code{(n x p)} with binary entries. Missing data should be coded as \code{NA}.}

\item{q}{Number of latent variables.}

\item{Qmatrix}{Q-matrix.}

\item{control}{List of control parameters (see 'Details').}

\item{start.par}{(optional) For simulation use. List of size 4 with starting model parameters (A, C, D, R).}

\item{...}{Further arguments to be passed to \code{control}.}
}
\value{
A list with components:
\itemize{
\item \code{G}: A matrix \code{(p x (q+1))} of estimated parameters.
\item \code{mu}: An vector \code{(q x 1)} of estimated latent variable means.
\item \code{R}: A matrix \code{(q x q)} of estimated correlations for the latent variables
\item \code{llk}: Marginal log-likelihood (computed via Monte-Carlo integration) evaluated at \code{A}, \code{C}, and \code{R}.
\item \code{AIC}: Akaike information criterion for the estimated model.
\item \code{BIC}: Bayesian (Schwarz) information criterion for the estimated model.
\item \code{cdllk.trace}: (if return.trace = T) A matrix \code{(iter x 2)} with the trace for the observed variables and latent variables log-likelihood.
\item \code{ar.trace}: (if return.trace = T) A vector \code{(iter)} with the trace for the acceptance rate (for MALA and RWMH samplers).
\item \code{theta.trace}: (if return.trace = T) A matrix \code{(iter x dim(theta))} with the trace for the parameter estimates.
}
}
\description{
Fit an Partial-Mastery Additive Cognitive Diagnosis Model (PM-CDM).
}
\details{
Test
}
\author{
Camilo Cárdenas-Hurtado (\email{c.a.cardenas-hurtado@lse.ac.uk}).
}
