% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmCDM.R
\name{apmCDM_findqCV}
\alias{apmCDM_findqCV}
\title{Find number of latent variables (q) using cross-validation (PM-ACDM)}
\usage{
apmCDM_findqCV(Ytrain, Ytest, q, control = list(), start.par = NULL, ...)
}
\arguments{
\item{Ytrain}{Matrix with observed binary entries to train the model. Missing entries (\code{Ytest}) coded as \code{NA}.}

\item{Ytest}{Matrix with observed binary entries to test the model. All entries are \code{NA} but the missing in \code{Ytrain}.}

\item{q}{Number of latent variables.}

\item{control}{List of control parameters (see 'Details').}

\item{...}{Further arguments to be passed to \code{control}.}
}
\value{
A list with components:
\itemize{
\item \code{CV.error}: Cross-validation error.
\item \code{AUC}: Area under the curve error.
\item \code{llk.gapmCDM} Marginal log-likelihood for the gapmCDM model
}
}
\description{
Find number of latent variables (q) using cross-validation (PM-ACDM)
}
\details{
Define CV.error.
}
\author{
Camilo Cárdenas-Hurtado (\email{c.a.cardenas-hurtado@lse.ac.uk}).
}
