% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmCDM.R
\name{apmCDM_mllkCV}
\alias{apmCDM_mllkCV}
\title{Find number of latent variables (q) using cross-validated MLLK (PM-ACDM)}
\usage{
apmCDM_mllkCV(Ytrain, Ytest, q, control = list(), ...)
}
\arguments{
\item{Ytrain}{Matrix with observed binary entries to train the model. Missing entries (\code{Ytest}) coded as \code{NA}.}

\item{Ytest}{Matrix with observed binary entries to test the model. All entries are \code{NA} but the missing in \code{Ytrain}.}

\item{q}{Number of latent variables.}

\item{control}{List of control parameters (see 'Details').}

\item{...}{Further arguments to be passed to \code{control}.}
}
\value{
A list with components:
\itemize{
\item \code{mllk.test}: Test data marginal log-likelihood.
\item \code{mllk.train}: Train data marginal log-likelihood.
\item \code{mllk.BIC} Train data BIC.
}
}
\description{
Find number of latent variables (q) using cross-validated MLLK (PM-ACDM)
}
\details{
Define CV.error.
}
\author{
Camilo Cárdenas-Hurtado (\email{c.a.cardenas-hurtado@lse.ac.uk}).
}
