% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmCDM.R
\name{apmCDM_sim}
\alias{apmCDM_sim}
\title{Simulate data from a Generalized Additive Partial Mastery Cognitive Diagnosis Model (GaPM-CDM).}
\usage{
apmCDM_sim(n, p, q, Qmatrix, control = list(), start.par = NULL, ...)
}
\arguments{
\item{n}{Number of simulated entries.}

\item{p}{Number of observed (binary) variables.}

\item{q}{Number of latent variables.}

\item{Qmatrix}{Q-matrix}

\item{control}{List of control parameters (see 'Details').}

\item{start.par}{(optional) List of size 3 with starting model parameters (G, mu, R).}

\item{...}{Further arguments to be passed to \code{control}.}
}
\value{
A list with components:
\itemize{
\item \code{Y}: A matrix \code{(n x p)} of simulated observed (binary) variables.
\item \code{Z}: A matrix \code{(n x q)} of simulated latent variables (on the continuous scale).
\item \code{U}: A matrix \code{(n x q)} of simulated latent variables (on the \link{0,1} scale).
\item \code{PI}: A matrix \code{(n x p)} of predicted probabilities.
}
}
\description{
Simulate data from a Generalized Additive Partial Mastery Cognitive Diagnosis Model (GaPM-CDM).
}
\details{
Test
}
\author{
Camilo Cárdenas-Hurtado (\email{c.a.cardenas-hurtado@lse.ac.uk}).
}
