% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmCDM.R
\name{gapmCDM_findqUI}
\alias{gapmCDM_findqUI}
\title{Find number of latent variables (q) using Universal inference}
\usage{
gapmCDM_findqUI(
  data,
  qmax = 10,
  qmin = 2,
  alpha = 0.05,
  type = "cross-fit",
  control = list(),
  ...
)
}
\arguments{
\item{data}{Matrix \code{(n x p)} with binary entries. Missing data should be coded as \code{NA}.}

\item{qmax}{Maximum number of latent variables to test.}

\item{qmin}{Minimum number of latent variables to test.}

\item{alpha}{Critical value for the universal inference test.}

\item{type}{Type of test for Universal inference.}

\item{control}{List of control parameters (see 'Details').}

\item{...}{Further arguments to be passed to \code{control}.}
}
\value{
A list with components:
\itemize{
\item \code{q}: Number of latent variables when Ho: dim(Z) = q is not rejected.
}
}
\description{
Find number of latent variables (q) using Universal inference
}
\details{
Test
}
\author{
Camilo Cárdenas-Hurtado (\email{c.a.cardenas-hurtado@lse.ac.uk}).
}
