% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logLik.pmCDM}
\alias{logLik.pmCDM}
\title{Compute marginal log-likelihood (via Importance Sampling) for an object of class "pmCDM".}
\usage{
\method{logLik}{pmCDM}(mod, Y, control = list(), ...)
}
\arguments{
\item{mod}{\code{pmCDM} object.}
}
\value{
A list with components:
\itemize{
\item \code{llk}: Marginal log-likelihood (computed via Monte-Carlo integration) evaluated at \code{A}, \code{C}, and \code{R}.
}
}
\description{
Compute marginal log-likelihood (via Importance Sampling) for an object of class "pmCDM".
}
\details{
Test
}
\author{
Camilo Cárdenas-Hurtado (\email{c.a.cardenas-hurtado@lse.ac.uk}).
}
