% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmCDM-package.R
\docType{package}
\name{pmCDM-package}
\alias{pmCDM}
\alias{pmCDM-package}
\title{pmCDM: Partial Mastery Cognitive Diagnosis Models}
\description{
Functions for fitting and simulating Partial Mastery Cognitive Diagnosis Models.
}
\author{
\strong{Maintainer}: Camilo Cárdenas-Hurtado \email{c.a.cardenas-hurtado@lse.ac.uk} (\href{https://orcid.org/0009-0003-1073-0912}{ORCID})

}
\keyword{internal}
