% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmCDM.R
\name{pmCDM.CV.error}
\alias{pmCDM.CV.error}
\title{Compare GaPM-CDM vs. aPM-CDM via negative cross-entropy error.}
\usage{
pmCDM.CV.error(
  Ytrain,
  Ytest,
  Qmatrix,
  q,
  controlA = list(),
  controlB = list(),
  ...
)
}
\arguments{
\item{Ytrain}{Matrix with observed binary entries to train the model. Missing entries (\code{Ytest}) coded as \code{NA}.}

\item{Ytest}{Matrix with observed binary entries to test the model. All entries are \code{NA} but the missing in \code{Ytrain}.}

\item{Qmatrix}{Q-matrix for aPM-CDM.}

\item{q}{Number of latent variables.}

\item{controlA}{List of control parameters for GaPM-CDM (see 'Details').}

\item{controlB}{List of control parameters for aPM-CDM (see 'Details').}

\item{...}{Further arguments to be passed to \code{controlA} and \code{controlB}.}
}
\value{
A list with components:
\itemize{
\item \code{CV.error.A}: Cross-validation error for GaPM-CDM.
\item \code{AUC.A}: Area under the curve error for GaPM-CDM.
\item \code{CV.error.B}: Cross-validation error for aPM-CDM.
\item \code{AUC.B}: Area under the curve error for aPM-CDM.
}
}
\description{
Compare GaPM-CDM vs. aPM-CDM via negative cross-entropy error.
}
\details{
Define CV.error.
}
\author{
Camilo Cárdenas-Hurtado (\email{c.a.cardenas-hurtado@lse.ac.uk}).
}
